Add titles app and move models to it
