+++ api_yamdb/api/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from reviews.models import User, Genre, Category, Title
+from reviews.models import User
 
 
 @admin.register(User)
@@ -15,40 +15,4 @@ class UserAdmin(admin.ModelAdmin):
     )
     list_editable = ('role',)
     search_fields = ('username',)
-    empty_value_display = '-пусто-'
-
-
-@admin.register(Category)
-class CategoryAdmin(admin.ModelAdmin):
-    list_display = (
-        'name',
-        'slug',
-    )
-    search_fields = ('name',)
-    list_filter = ('name',)
-    empty_value_display = '-пусто-'
-
-
-@admin.register(Genre)
-class GenreAdmin(admin.ModelAdmin):
-    list_display = (
-        'name',
-        'slug',
-    )
-    search_fields = ('name',)
-    list_filter = ('name',)
-    empty_value_display = '-пусто-'
-
-
-@admin.register(Title)
-class TitleAdmin(admin.ModelAdmin):
-    list_display = (
-        'id',
-        'category',
-        'name',
-        'year',
-        'description'
-    )
-    search_fields = ('name',)
-    list_filter = ('year', 'category', 'genre')
     empty_value_display = '-пусто-'
