+++ api_yamdb/api_yamdb/settings.py
@@ -26,6 +26,7 @@ INSTALLED_APPS = [
     'rest_framework_simplejwt',
     'api',
     'reviews',
+    'titles',
     'django_filters'
 ]
