+++ api_yamdb/reviews/models.py
@@ -3,7 +3,7 @@ from django.core.validators import MaxValueValidator, MinValueValidator
 from django.db import models
 from django.utils import timezone
 
-from .validators import year_validator
+from titles.models import Title
 
 ROLES = [
     ('user', 'Пользователь'),
@@ -25,99 +25,6 @@ class User(AbstractUser):
                             )
     bio = models.TextField(verbose_name='Биография', blank=True)
     updated_at = models.DateTimeField(default=timezone.now)
-
-
-class Category(models.Model):
-    name = models.CharField(
-        'Название категории',
-        max_length=256
-    )
-    slug = models.SlugField(
-        'Слаг категории',
-        max_length=50,
-        db_index=True,
-        unique=True
-    )
-
-    class Meta:
-        verbose_name = 'Категория'
-        verbose_name_plural = 'Категории'
-
-    def __str__(self) -> str:
-        return self.slug
-
-
-class Genre(models.Model):
-    name = models.CharField(
-        'Название жанра',
-        max_length=256
-    )
-    slug = models.SlugField(
-        'Слаг жанра',
-        max_length=50,
-        db_index=True,
-        unique=True
-    )
-
-    class Meta:
-        verbose_name = 'Жанр'
-        verbose_name_plural = 'Жанры'
-
-    def __str__(self) -> str:
-        return self.slug
-
-
-class Title(models.Model):
-    name = models.CharField(
-        'Название произведения',
-        max_length=256,
-        db_index=True
-    )
-    year = models.IntegerField(
-        'Год выпуска',
-        validators=[year_validator]
-    )
-    description = models.TextField(
-        'Описание',
-        null=True,
-        blank=True
-    )
-    genre = models.ManyToManyField(
-        Genre,
-        verbose_name='Жанр произведения',
-        blank=False,
-        related_name='titles',
-
-    )
-    category = models.ForeignKey(
-        Category,
-        verbose_name='Категория произведения',
-        on_delete=models.SET_NULL,
-        null=True,
-        blank=False,
-        related_name='titles'
-    )
-
-    class Meta:
-        verbose_name = 'Произведение'
-        verbose_name_plural = 'Произведения'
-
-    def __str__(self):
-        return self.name
-
-
-class GenreTitle(models.Model):
-    title = models.ForeignKey(
-        Title,
-        on_delete=models.CASCADE,
-    )
-    genre = models.ForeignKey(
-        Genre,
-        on_delete=models.CASCADE,
-    )
-
-    def __str__(self) -> str:
-        return f'{self.title} {self.genre}'
 
 
 class Review(models.Model):
