+++ api_yamdb/titles/admin.py
@@ -1,0 +1,39 @@
+from django.contrib import admin
+
+from titles.models import Category, Title, Genre
+
+
+@admin.register(Category)
+class CategoryAdmin(admin.ModelAdmin):
+    list_display = (
+        'name',
+        'slug',
+    )
+    search_fields = ('name',)
+    list_filter = ('name',)
+    empty_value_display = '-пусто-'
+
+
+@admin.register(Genre)
+class GenreAdmin(admin.ModelAdmin):
+    list_display = (
+        'name',
+        'slug',
+    )
+    search_fields = ('name',)
+    list_filter = ('name',)
+    empty_value_display = '-пусто-'
+
+
+@admin.register(Title)
+class TitleAdmin(admin.ModelAdmin):
+    list_display = (
+        'id',
+        'category',
+        'name',
+        'year',
+        'description'
+    )
+    search_fields = ('name',)
+    list_filter = ('year', 'category', 'genre')
+    empty_value_display = '-пусто-'
